+++ api/routes/parents.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const { getAllParents, createParent } = require('../controllers/parentController');
+const { getAllParents, createParent, getParentById } = require('../controllers/parentController');
 
 router.get('/', getAllParents);
 router.post('/', createParent);
+router.get('/:id', getParentById);
 
 module.exports = router;